+++ 3rdparty/kshortestpaths/Graph.h
@@ -49,16 +49,16 @@ protected: // members
 	// Basic information
 	map<BaseVertex*, set<BaseVertex*>*> m_mpFanoutVertices;
 	map<BaseVertex*, set<BaseVertex*>*> m_mpFaninVertices;
-	map<int, double> m_mpEdgeCodeWeight; 
+	map<long, double> m_mpEdgeCodeWeight;
 	vector<BaseVertex*> m_vtVertices;
-	int m_nEdgeNum;
-	int m_nVertexNum;
+	long m_nEdgeNum;
+	long m_nVertexNum;
 
-	map<int, BaseVertex*> m_mpVertexIndex;
+	map<long, BaseVertex*> m_mpVertexIndex;
 
 	// Members for graph modification
-	set<int> m_stRemovedVertexIds;
-	set<pair<int,int> > m_stRemovedEdge;
+	set<long> m_stRemovedVertexIds;
+	set<pair<long,long> > m_stRemovedEdge;
 
 public:
 
@@ -70,13 +70,13 @@ public:
 
 	void clear();
 
-	void add_vertex(int node_id);
-	void add_vertex(int node_id,double h);
-	void add_edge(int start_vertex, int end_vertex, double edge_weight);
+	void add_vertex(long node_id);
+	void add_vertex(long node_id,double h);
+	void add_edge(long start_vertex, long end_vertex, double edge_weight);
 
-	BaseVertex* get_vertex(int node_id);
+	BaseVertex* get_vertex(long node_id);
 	
-	int get_edge_code(const BaseVertex* start_vertex_pt, const BaseVertex* end_vertex_pt) const;
+	long get_edge_code(const BaseVertex* start_vertex_pt, const BaseVertex* end_vertex_pt) const;
 	set<BaseVertex*>* get_vertex_set_pt(BaseVertex* vertex_, map<BaseVertex*, set<BaseVertex*>*>& vertex_container_index);
 
 	double get_original_edge_weight(const BaseVertex* source, const BaseVertex* sink);
@@ -86,12 +86,12 @@ public:
 	void get_precedent_vertices(BaseVertex* vertex, set<BaseVertex*>& vertex_set);
 
 	/// Methods for changing graph
-	void remove_edge(const pair<int,int> edge)
+	void remove_edge(const pair<long,long> edge)
 	{
 		m_stRemovedEdge.insert(edge);
 	}
 
-	void remove_vertex(const int vertex_id)
+	void remove_vertex(const long vertex_id)
 	{
 		m_stRemovedVertexIds.insert(vertex_id);
 	}
@@ -106,12 +106,12 @@ public:
 		m_stRemovedVertexIds.clear();
 	}
 
-	void recover_removed_edge(const pair<int,int> edge)
+	void recover_removed_edge(const pair<long,long> edge)
 	{
 		m_stRemovedEdge.erase(m_stRemovedEdge.find(edge));
 	}
 
-	void recover_removed_vertex(int vertex_id)
+	void recover_removed_vertex(long vertex_id)
 	{
 		m_stRemovedVertexIds.erase(m_stRemovedVertexIds.find(vertex_id));
 	}
